Safari-609.1.11.1/LayoutTests/inspector/worker/dom-debugger-dom-breakpoints-expected.txt

Tests for DOM breakpoint functionality (DOMDebugger) in a Worker.


== Running test suite: Worker.DOMDebugger.DOMBreakpoint
-- Running test case: Worker.DOMDebugger.DOMBreakpoint.setDOMBreakpoint
PASS: Should produce an exception.
Error: Not supported

-- Running test case: Worker.DOMDebugger.DOMBreakpoint.removeDOMBreakpoint
PASS: Should produce an exception.
Error: Not supported
